* Removed redundant cast * Surpressed odd Warning in a Test

+++ app/src/main/java/de/dhbwloerrach/beaconlocation/models/DistanceList.java
@@ -25,6 +25,6 @@ public class DistanceList extends ArrayList<TimedDistance> {
         for (TimedDistance distance : this)
             sum += distance.getDistance();
 
-        return ((double) sum) / this.size();
+        return sum / this.size();
     }
 }
